requirements.txt
Flask>=2.3
Flask_SQLAlchemy>=3.0
flask-jwt-extended>=4.7
passlib==1.7.4
argon2-cffi>=21.0.0
python-dotenv>=1.0
PyMySQL>=1.1
